#59 - Add admin panel
