+++ src/main/java/com/company/simulator/model/User.java
@@ -59,6 +59,10 @@ public class User implements UserDetails {
         return roles.contains(Role.ADMIN);
     }
 
+    public boolean isTeacher() {
+        return roles.contains(Role.TEACHER);
+    }
+
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         return getRoles();
