+++ src/main/resources/templates/admin/common.ftlh
