+++ src/main/resources/templates/admin/userList.ftlh
@@ -1,20 +1,19 @@
-<#import "parts/common.ftlh" as c>
+<#import "../parts/common.ftlh" as c>
 
 <@c.page>
-    List of users
-
-    <table>
+    <h5>List of users</h5>
+    <table class="table table-bordered" style="width: auto">
         <tr>
             <th>Name</th>
             <th>Role</th>
-            <th></th>
+            <th>Editing</th>
         </tr>
         <tbody>
         <#list users as user>
             <tr>
                 <td>${user.username!"null or missing"}</td>
                 <td><#list user.roles as role>${role}<#sep>, </#list></td>
-                <td><a href="/user/${user.id}">Edit</a></td>
+                <td><a class="btn btn-primary" href="/admin/user/${user.id}">Edit</a></td>
             </tr>
         </#list>
         </tbody>
