+++ src/main/resources/templates/parts/head.ftlh
@@ -1,0 +1,12 @@
+<head>
+    <meta charset=UTF-8"/>
+    <title>SQL simulator</title>
+    <link rel="stylesheet" href="/static/style.css">
+
+    <!-- Required meta tags -->
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
+          integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
+          crossorigin="anonymous">
+</head>
