+++ src/main/resources/templates/parts/security.ftlh
@@ -6,12 +6,14 @@ known = Session.SPRING_SECURITY_CONTEXT??
     user = Session.SPRING_SECURITY_CONTEXT.authentication.principal
     name = user.getUsername()
     isAdmin = user.isAdmin()
+    isTeacher = user.isTeacher()
     currentUserId = user.getId()
     >
 <#else>
     <#assign
     name = "unknown"
     isAdmin = false
+    isTeacher = false
     currentUserId = -1
     >
 </#if>
